requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Template Engine
Jinja2==3.1.2

# Database & ORM
SQLAlchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Supabase
supabase==2.3.0
postgrest==0.13.0

# Forms & Validation
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Authentication & Security
Flask-Login==0.6.3
itsdangerous==2.1.2
bcrypt==4.1.2

# Rate Limiting
Flask-Limiter==3.5.0

# Email
Flask-Mail==0.9.1

# HTTP Client
requests==2.31.0
httpx==0.25.2

# Image Processing
Pillow==10.1.0

# Environment Variables
python-dotenv==1.0.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
click==8.1.7
python-slugify==8.0.1

# Excel/CSV Processing
openpyxl==3.1.2
pandas==2.1.4

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
playwright==1.40.0
faker==21.0.0

# Code Quality
flake8==6.1.0
black==23.12.1
ruff==0.1.8

# Monitoring (optional)
sentry-sdk[flask]==1.39.1
